refactor: complete migration to new MarketsTable structure
- Update imports to use new MarketsTable path
- Remove old MarketsTableWithSameLoanAsset.tsx file
- Fix import paths in MarketSelectionOnboarding and MarketSelectionModal

All components now use the modularized MarketsTable structure with Zustand state management.
